Review feedback refactoring
